Start porting `old-parser` test case inputs; introduce (too much) conversion plumbing.

+++ primval/src/parseutil.rs
@@ -3,7 +3,7 @@ use chumsky::{text, Parser as _};
 use sappho_syntax_parsable::error::ChumskyError;
 use sappho_syntax_parsable::Parser;
 
-pub(crate) fn number() -> impl Parser<i64> {
+pub(crate) fn number() -> impl Parser<i32> {
     let disallowed_trailing_char = filter(|&c: &char| c.is_alphabetic() || c.is_control())
         .try_map(|c, span| -> Result<(), ChumskyError> {
             Err(ChumskyError::custom(
@@ -16,7 +16,7 @@ pub(crate) fn number() -> impl Parser<i64> {
     text::digits(10)
         .then_ignore(disallowed_trailing_char)
         .try_map(|digs: String, span| {
-            <i64 as std::str::FromStr>::from_str(&digs)
+            <i32 as std::str::FromStr>::from_str(&digs)
                 .map_err(|e| ChumskyError::custom(span, e.to_string()))
         })
         .labelled("number")
